LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/mpadded/mpadded-percentage-002-expected.txt


PASS width='200%' is interpreted as the default value
PASS height='200%' is interpreted as the default value
PASS depth='200%' is interpreted as the default value
PASS lspace='200%' is interpreted as the default value
PASS voffset='200%' is interpreted as the default value
